+++ Portierniaktosiedzi.Tests/Portierniaktosiedzi.Tests.csproj
@@ -55,7 +55,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Models\HolidaysShould.cs" />
+    <Compile Include="Utility\HolidaysShould.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
