+++ Portierniaktosiedzi/Portierniaktosiedzi.csproj
@@ -75,7 +75,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Models\Employee.cs" />
-    <Compile Include="Models\Holidays.cs" />
+    <Compile Include="Utility\Holidays.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
